finished clustering and sorting model
this is a model for clustering when call through API, first its get the user data from SQL then the model will do clustering, after getting the clustering assignment. it will update the database with new data and then it take the drivers data and the user data that already been clustered and then it will sort the data in order to assigned driver then it will use tsp greedy to make a routing in order.